refactor inform agreed extension date
